Factor out schedule-metadata.js.
