+++ frontend/source/js/data-explorer/tests/schedule.test.jsx
@@ -2,7 +2,7 @@ import toJson from 'enzyme-to-json';
 
 import { Schedule } from '../components/schedule';
 import makeSetup from './testSetup';
-import { SCHEDULE_LABELS } from '../constants';
+import { scheduleLabels } from '../schedule-metadata';
 
 const defaultProps = {
   schedule: '',
@@ -19,9 +19,9 @@ describe('<Schedule>', () => {
     expect(select.exists()).toBeTruthy();
     const options = wrapper.find('option');
     // should have options for all schedules + 1 for "(all)"
-    expect(options.length).toBe(Object.keys(SCHEDULE_LABELS).length + 1);
-    Object.keys(SCHEDULE_LABELS).forEach((sched) => {
-      const title = SCHEDULE_LABELS[sched];
+    expect(options.length).toBe(Object.keys(scheduleLabels).length + 1);
+    Object.keys(scheduleLabels).forEach((sched) => {
+      const title = scheduleLabels[sched];
       const opt = wrapper.find(`option[value="${sched}"]`);
       expect(opt.exists()).toBeTruthy();
       expect(opt.text()).toBe(title);
